use .classList.toggle instead of add / remove
if toggle is undefined, `.toggle` will act as a flip flop (wich the name "toggle" should imply)
if toggle is true, it will add the class unless it is already contained
if toggle is false, it will remove the class unless it is already removed

alternatively i'd probably suggest renaming it from toggle_class to something else. a toggle is a state aware flip flop by default.

any objections?