Don't use shorthand binding in textarea example
The example uses a variable called `value` and a shorthand binding notation `<textarea bind:value>`. I think this example is simpler, and more consistent with other binding examples, if it doesn't rely on the reader knowing about the shorthand notation. For example, I was confused when I renamed the `value` variable to `text`, and it didn't work.

```js
<script>
	let text = `Some words are *italic*, some are **bold**`;
</script>
<textarea bind:text></textarea>
```

This commit uses a variable named `text`, rather than `value`, and avoids the shorthand binding form.