more explanations
the unsubscribe needs to be explained here. In the API docs it says "will be unsubscribed when appropriate" when using $ store prefix, but it's not clear what does "appropriate" means.   
The API full note: "Any time you have a reference to a store, you can access its value inside a component by prefixing it with the $ character. This causes Svelte to declare the prefixed variable, and set up a store subscription that will be unsubscribed when appropriate."