Update old `.technology` links in the Sapper intro blog post
`sapper.svelte.technology` has moved to `sapper.svelte.dev`. The old
URLs redirect to the `.dev` site, but updating the links will avoid the
need for each browser to follow a redirect before getting to the site.
Additionally, the fragment id in the link to the description of the
Sapper name has changed from `#why-the-name-` to `#Why_the_name`. I've
updated that as well, so that following the link will scroll to the
appropriate section of the page.