Move function declaration before reference
The way the example is currently presented requires that the user have an understanding of how Javascript hoists functions. By moving the function declaration before the function reference, the order of the code becomes more clear to the reader, which I think is valuable here. I think it's more important that this material clearly teach the behavior of Svelte, rather than the nuances of Javascript.