More comprehensible Stores type
If Stores can be a "Readable<any>" or a bunch of "Readable<any>"s why not make that array a type that is better readable (Readable<any>[]). Otherwise I have to parse an Array of Readables like that: "list as [Readable<any>, ...Array<Readable<any>>]" which is not very nice.