Ensuring parentNode exists before detaching
This fix adds a check to prevent error in case the parentNode is null. This error is encountered in situations where raw @html content is used and has been mutated by non-svelte code. In some cases, when the raw @html code is being detached from the DOM, it is no longer part of the DOM.