[store] Make writable depend on readable
...and add get function to stores.

The first part came along trying to fix the second one. I just noticed how it was and decided the fix was easy enough. 
I did not run any test and actually coded that inside github's default editor, and it's my first contribution to Svelte, so please excuse me if this PR is completely useless/stupid.

Reason for the PR: 
The current `get(store)` access function is (inefficient)[https://svelte.dev/docs#get] (see the warning) and not object oriented.

Fix:
The easy fix is to add a `get` accessor for the readable's value.