Clarify that VS Code isn't the only supported IDE
First time I ever looked through the tutorial the wording here made it seem like the only way to get a great IDE/Svelte tooling experience was with VS Code. Because of that I just didn't bother learning Svelte at the time.

Always good to be as inclusive as possible!