store subscribers shouldn't get the initial undefined value
before this, an empty, uninitialized store would emit an `undefined`
value, even before any value is set. this is quite unexpected and not
the same behaviour as rxjs (for example).

you can still explicitly create a store with undefined with
`writeable(undefined)` or `readable(undefined)`

Previously:
```
> var store = require('./store');
> var w = store.writable();
> w.subscribe(v=>console.log('v:', v));
v: undefined
```

After this commit:
```
> var store = require('./store');
> var w = store.writable();
> w.subscribe(v=>console.log('v:', v));
> w.set('test');
v: test
```