Improve clarity of 2 files to edit and improve learning
While going through the tutorial, I didn't notice the 2nd file to edit in the tutorial or the way that Svelte directly uses a component and by this point in the tutorial had forgot what exactly components themselves are at this stage.  Repetition is the key to reinforcement learning, so let's repeat what components are at this critical point of learning.
This fixes all 3 issues and importantly reaffirms what components are.