Add comment about generating a Svelte project without Kit
We're seeing many of our students trip up with Svelte Docs generating applications that require NodeJS on the server. Turning a SvelteKit project into a Svelte-only project is not trivial.

This PR adds a line in the "Getting Started" section about how to generate frontend-only sites/apps while using NodeJS and Svelte for local development.