Update text.md
I am not a good developer, and I spent a long time trying to figure out the difference between the code with and without the bind. It probably seems clear that I should've figured it out, but adding a statement to show how to invoke the changes might help another person like me. A prompt to type in the DOM could help somebody more quickly pass this level.