Mention the interactivity of the installer
Having never installed Svelte before today, I was waiting for a long time at the initial prompt for it to... do something.  It never did.  Not because it was broken, but because the fact I needed to use my arrow keys was not obvious at all.  It asked what template I wanted to use, but didn't explain how to make the selection.  Once I figured it out (after asking in Discord) it was blatantly obvious, but for a very first time user and after having installed other similar products that don't use this type of interactivity but start with effectively identical commands, it was anything but obvious.  Almost made me cross Svelte off my evaluation list.