Hint matches existing structure and solution
In the discussion there's a hint-- a bit of code that, when added to the existing code, gives the solution to the problem. The bit of code in the hint does not match the pre-existing code or the solution in these ways:
1) different variable name-- number vs. value
2) what transition to use
3) whether to use  the transition attribute or the in attribute
I have just changed the markup so that the suggested code matches what needs to be added to most quickly arrive at the solution. Otherwise the learner has to also fix/grapple with these other differences.