fix: don't freeze non-POJOs
The main reason for Object.freeze on `$state.frozen` is to mark a boundary for `$state` - but since that one doesn't traverse non-POJOs either, we should leave class instances etc alone, too.
Since "this is a POJO and not a class instance" cannot be described in typescript, the `Readonly` modifier had to be removed from the return type.