chore: failed attempt for binding+events order
This is attempt 1 to fix #11138. The idea is that actions, bindings and event directives are added to the "spread attributes" function. That function then first sets all attributes, collects event attributes, then runs actions/bindings/event handlers, and then runs event attributes, to guarantee that event attributes run last.

The sad thing is that it doesn't work. The problem is that bindings or actions could have effects inside them, and those effects would then be tied to the render effect surrounding the "spread attributes" function, and not to the block effect above it, leading to wrong rerun/destroy timings.