fix: properly delay intro transitions (#12389)
* fix: properly delay intro transitions

WAAPI applies the styles of a delayed animation only when that animation starts. In the case of fade-in transitions that means the element is visible, then goes invisible and fades in. Fix that by never applying a delay on intro transitions, instead add keyframes of the initial state for the duration of the delay.

Fixes #10876

* fix bug, make test pass

* make test more selfcontained, test outro delay aswell and add functionality for that in animation-helpers

* lint

---------

Co-authored-by: Rich Harris <rich.harris@vercel.com>