fix: ensure types are easier to follow for TypeScript
As a byproduct of #12239 some types were changed from being directly exported to indirectly exported. TypeScript is bad at following these types while generating d.ts files, so we can't do that. A check at the end of the types generation was added to prevent regressions in the future. Fixes #13128

Also removes the old types from `svelte/reactivity` which are deprecated for a while now.